refactor: move page routing out of Nav component and use useLocation hook to check path (#53)
